web/fast-web/fast-proj1/pyproject.toml
[project]
name = "fast-proj1"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.135.1",
    "psycopg2>=2.9.11",
    "pyjwt>=2.11.0",
    "python-decouple>=3.8",
    "routes>=2.5.1",
    "sqlmodel>=0.0.37",
    "uvicorn>=0.41.0",
]
